Possibility to lock UI instead of warning
By adding SESSION_SECURITY_WARN_LOCK=True to settings it is possible to
lock the UI at the warning timeout so that the user must input password
to unlock UI instead of e.g. moving mouse. If user tries to browse to
another page when UI is locked, the session is immediately logged out
by the middleware.